Template must be ALL in .h
.cpp not available at compile time
